Better document show base types
