+++ pyramidtriangles/shows/fuck_your_burn.py
@@ -6,7 +6,7 @@ from ..color import HSV
 from ..grid import Position, Pyramid
 from ..ponzicolor import Color, space
 from ..randomcolor import random_color
-from .show import Show, ValueKnob, HSVKnob
+from . import Show, ValueKnob, HSVKnob
 
 
 class FuckYourBurn(Show):
