+++ pyramidtriangles/shows/random_cells.py
@@ -1,7 +1,7 @@
 from collections import deque
 import random
 
-from .show import Show
+from . import Show
 from ..color import RGB
 from ..grid import Cell, Grid, Pyramid
